fix slow highlighter on vim 7.4
Uses "syntax keyword" instead of "keyword match" to find subjects to
highlight, which is much faster. The problem is that the type of regex used
by vim-easytags is a pathological bad case for the NFA regex engine which
was introduced in vim 7.4.

As reported here:
https://groups.google.com/forum/#!topic/vim_dev/cPcMap1BdQw

The patch was taken from the vim-easytags issues list on github, credit for
it goes to @juliantaylor.